LayoutTests/media/video-source-expected.txt

EVENT(loadstart)
EXPECTED (relativeURL(video.currentSrc) == 'content/test.mp4') OK
EXPECTED (source.getAttribute('src') == 'content/test.mp4') OK
EXPECTED (relativeURL(source.src) == 'content/test.mp4') OK
END OF TEST
